+++ editor/settings/editor_settings.h
@@ -154,6 +154,7 @@ public:
 	Variant get_setting(const String &p_setting) const;
 	bool has_setting(const String &p_setting) const;
 	void erase(const String &p_setting);
+	void erase_group(const String &p_group);
 	void raise_order(const String &p_setting);
 	void set_initial_value(const StringName &p_setting, const Variant &p_value, bool p_update_current = false);
 	void set_restart_if_changed(const StringName &p_setting, bool p_restart);